feat: added conditional for the navbar based on the user role.
This commit adds changes to the navbar to display only the sections that the user has permissions to based on their role.
