+++ frontend/src/app/dashboard/page.tsx
@@ -1,9 +1,25 @@
-import Dashboard from "@/components/dashboard/Dashboard";
+"use client";
 
-export default function page() {
+import { AuthenticationContext } from "@/context/AuthContext";
+import { useContext } from "react";
+import Navbar from "../../components/navbar/navbar";
+
+export default function Dashboard() {
+    const authContext = useContext(AuthenticationContext);
+
+    if (!authContext) {
+        return <div>loading</div>;
+    }
+
+    const { isLoggedIn, user, logout } = authContext;
+
+    console.log("isLoggedIn: " + isLoggedIn);
+    console.log("User: " + user);
+
     return (
-        <div>
-            <Dashboard></Dashboard>
-        </div>
+        <>
+            <Navbar></Navbar>
+            <div>User role: {user?.role}</div>
+        </>
     );
 }
