+++ frontend/src/app/page.tsx
@@ -1,16 +1,22 @@
-import { SignIn } from "@/components/authentication/SignIn";
+"use client";
+
 import { SignUp } from "@/components/authentication/SignUp";
-import Navbar from "@/components/navbar/navbar";
+import { SignIn } from "@/components/authentication/SignIn";
+import { useState } from "react";
 
 export default function Home() {
+    const [showSignUp, setShowSignUp] = useState(false);
+
     return (
         <div>
-            <Navbar />
             <main>
-                <h1>Content Goes Here</h1>
+                <h1>SPEED</h1>
+                <h3>Please Log in to Continue</h3>
+                {showSignUp ? <SignUp /> : <SignIn />}
+                <button onClick={() => setShowSignUp(!showSignUp)}>
+                    {showSignUp ? "Go to Login" : "Go to Sign Up"}
+                </button>
             </main>
-            <SignUp />
-            <SignIn />
         </div>
     );
 }
