--- frontend/src/components/dashboard/dashboard.tsx
@@ -1,25 +1,0 @@
-"use client";
-
-import { AuthenticationContext } from "@/context/AuthContext";
-import { useContext } from "react";
-import Navbar from "../navbar/navbar";
-
-export default function Dashboard() {
-    const authContext = useContext(AuthenticationContext);
-
-    if (!authContext) {
-        return <div>loading</div>;
-    }
-
-    const { isLoggedIn, user, logout } = authContext;
-
-    console.log("isLoggedIn: " + isLoggedIn);
-    console.log("User: " + user);
-
-    return (
-        <>
-            <Navbar></Navbar>
-            <div>User role: {user?.role}</div>
-        </>
-    );
-}
